requirements.txt
crispy-bootstrap5
Django
django-crispy-forms
django-bootstrap-v5
django-filter
djangorestframework
openpyxl
pandas
Pillow
psycopg2
